requirements.txt
dj-database-url==0.5.0
Django==1.11.11
djangorestframework==3.7.7
drf-extensions==0.3.1
psycopg2==2.7.4
pytz==2018.3
